requirements.txt
fastapi
uvicorn[standard]
paddleocr
paddlepaddle
python-multipart
pillow
numpy
opencv-python
pdfplumber
python-docx
python-multipart
